avcodec/jni: use size_t to store structure offsets

+++ libavcodec/ffjni.h
@@ -24,6 +24,7 @@
 #define AVCODEC_FFJNI_H
 
 #include <jni.h>
+#include <stddef.h>
 
 /*
  * Attach permanently a JNI environment to the current thread and retrieve it.
@@ -105,7 +106,7 @@ struct FFJniField {
     const char *method;
     const char *signature;
     enum FFJniFieldType type;
-    int offset;
+    size_t offset;
     int mandatory;
 
 };
